requirements.txt
pyyaml
python-binance
reportlab
pandas
numpy